handle plain string search identical to fulltext search
Example: A user enters a ZWJ emoji sequence, the result is a list of
the codepoints the emoji is composed of. Before this commit the user
was greeted by a large "no cps found" header. But that is the result
of an implementation detail that is of no interest to the user.